backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-allauth==0.57.0
django-cors-headers==4.2.0
django-filter==23.3
django-extensions==3.2.3
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.11
Pillow==12.1.1
